requirements.txt
Flask==0.12.2
Jinja2==2.9.6
Flask_SQLAlchemy==2.2
SQLAlchemy==1.1.13
gunicorn
newspaper3k
